requirements.txt
blinker==1.8.1
boto3==1.34.93
botocore==1.34.93
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
dash==2.16.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
et-xmlfile==1.1.0
flask==3.0.3
gunicorn==22.0.0
idna==3.7
importlib-metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
nest-asyncio==1.6.0
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
plotly==5.21.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.31.0
retrying==1.3.4
s3transfer==0.10.1
six==1.16.0
tenacity==8.2.3
typing-extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
werkzeug==3.0.2
zipp==3.18.1
